Move genserver logic outside Blocks API
